+++ packages/client/package.json
@@ -39,6 +39,7 @@
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "13.3.0",
     "@testing-library/user-event": "^14.4.3",
+    "@types/isomorphic-fetch": "0.0.36",
     "@types/jest": "28.1.8",
     "@types/lodash": "4.14.194",
     "@types/react": "18.0.17",
@@ -52,6 +53,7 @@
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-react": "7.32.2",
     "identity-obj-proxy": "3.0.0",
+    "isomorphic-fetch": "3.0.0",
     "jest": "28",
     "jest-canvas-mock": "2.5.1",
     "jest-environment-jsdom": "29.0.1",
